refactor(test): 외부 인터페이스 디렉토리 정리
